rabbit_plugins: Use `rabbit_misc:strict_version_minor_equivalent()` to check compatibility
That new function has the behavior that `version_minor_equivalent()` had
before: it considers 3.7.x and 3.8.x as incompatible. This is useful for
plugins because the feature flags subsystem only applies to inter-node
compatibility, not plugins.

Therefore a plugin marked as compatible with 3.7.x should not be
considered as compatible with 3.8.x.

[#159298729]